[WFLY-18120] Split LayersTestCase tests of wildfly-ee feature pack layers from tests of wildfly feature pack layers
