+++ testsuite/layers-expansion/full-ha-config.xml
@@ -1,0 +1,6 @@
+<!-- Allow to re-name standalone-full-ha.xml file standalone.xml, expected by tests -->    
+<config  xmlns="urn:jboss:galleon:config:1.0" model="standalone" name="standalone-full-ha.xml">
+    <props>
+        <prop name="--server-config" value="standalone.xml"/>
+    </props> 
+</config>
