+++ testsuite/layers-expansion/src/test/java/org/jboss/as/test/layers/expansion/LayersTestCase.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.as.test.layers.expansion;
+
+import org.jboss.as.test.shared.LayersTestBase;
+
+public class LayersTestCase extends LayersTestBase {
+    // Currently we get all functionality from the superclass; this class is just so surefire has something to run
+    // in this testsuite module
+}
