+++ testsuite/layers/src/test/java/org/jboss/as/test/layers/base/LayersTestCase.java
@@ -1,0 +1,40 @@
+/*
+ *
+ *  * JBoss, Home of Professional Open Source.
+ *  * Copyright 2023 Red Hat, Inc., and individual contributors
+ *  * as indicated by the @author tags.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *     http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+package org.jboss.as.test.layers.base;
+
+
+import org.jboss.as.test.shared.LayersTestBase;
+import org.jboss.as.test.shared.util.AssumeTestGroupUtil;
+
+public class LayersTestCase extends LayersTestBase {
+
+    @Override
+    public void test() throws Exception {
+        // For WildFly Preview testing, ignore this test. The testing in the layers-expansion
+        // testsuite module covers this.
+        // In this module the test-standalone-reference installation will be
+        // provisioned with a lot of MP, etc modules, as that's what wildfly-preview FP does.
+        // But the test-all-layers installation will not include the expansion layers as
+        // testing those is out of scope for this module.
+        AssumeTestGroupUtil.assumeNotWildFlyPreview();
+
+        super.test();
+    }
+}
